requirements.txt
# Core Framework
chainlit==2.8.3

# LLM and Embeddings
groq==0.30.0
langchain==0.3.7
langchain-groq==0.2.1
langchain-huggingface==0.1.2
langchain-community==0.3.5

# Embeddings Model
sentence-transformers==3.3.1

# Vector Store
faiss-cpu==1.9.0

# PDF Processing
pypdf==5.1.0

# Utilities
python-dotenv==1.0.1
tiktoken==0.8.0

# Optional: GPU acceleration for embeddings (if you have NVIDIA GPU)
# torch==2.5.1
# faiss-gpu==1.9.0